Troca exceção por mensagens objetivas ao remover paciente

+++ desafio-1.2/service/PacientesService.js
@@ -39,11 +39,17 @@ export class PacientesService {
   }
 
   excluirPeloCpf(cpf) {
+    if (!this.#pacientesRepository.existePacienteCadastradoComCpf(cpf)) {
+      console.log("Não existe um paciente cadastrado com o CPF informado.");
+      return false;
+    }
     if (this.#pacienteComConsultaAgendadaFutura(cpf)) {
-      throw "Paciente com uma consulta agendada futura não pode ser excluído."
+      console.log("Paciente com uma consulta agendada futura não pode ser excluído.");
+      return false;
     }
     this.#excluirConsultasPassadas(cpf);
     this.#pacientesRepository.excluir(cpf);
+    return true;
   }
 
   #pacienteComConsultaAgendadaFutura(cpf) {
